benchmarks/1_behavior_cls/rgb/requirements.txt
torch
torch-summary
torchvision
tqdm
numpy
pillow
scikit-learn
ultralytics
pytz
PyYAML
pandas
